Fix a couple of compiler warnings

+++ src/gui/C4DownloadDlg.cpp
@@ -125,11 +125,11 @@ void C4DownloadDlg::OnIdle()
 			// file size known: Download in progress
 			sStatus.Ref(LoadResStr("IDS_CTL_DL_PROGRESS"));
 			if (iSize <= 1024)
-				sSize.Format(" (%ld Bytes)", iSize);
+				sSize.Format(" (%ld Bytes)", (long)iSize);
 			else if (iSize <= 1024*1024)
-				sSize.Format(" (%ld KB)", iSize/1024);
+				sSize.Format(" (%ld KB)", (long)(iSize/1024));
 			else
-				sSize.Format(" (%ld MB)", iSize/1024/1024);
+				sSize.Format(" (%ld MB)", (long)(iSize/1024/1024));
 			iProgress = int64_t(100) * HTTPClient.getDownloadedSize() / iSize;
 		}
 	}
